Refactor : 모바일 헤더 로직을 커스텀 훅으로 분리
